conformance-suite/requirements.txt
PyYAML>=6.0
jsonschema>=4.0.0
click>=8.0.0